LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-pseudo-checkmark-expected.txt



PASS Resolution of width is correct when pseudo-element argument is ignored (due to no colon)
PASS Resolution of width is correct when pseudo-element argument is ignored (due to single-colon)
FAIL Resolution of width is correct for pseudo-element (due to double-colon) assert_equals: expected "13px" but got ""
